Factor NewtonianEuler/ComputeItems.?pp
We need some of the functions in a few different places, but not all of them. No
point in tying them all together.
